Move "lightness" checks to new file
