+++ skylight/src/main/java/drewhamilton/skylight/Lightness.kt
@@ -1,0 +1,42 @@
+@file:JvmName("SkylightLightness")
+package drewhamilton.skylight
+
+import java.time.ZonedDateTime
+
+//region Skylight
+/**
+ * Determine whether it is light outside at the given [coordinates] at the given [dateTime], where "light" means after
+ * dawn and before dusk on the given date.
+ */
+fun Skylight.isLight(coordinates: Coordinates, dateTime: ZonedDateTime) =
+    when (val skylightDay = getSkylightDay(coordinates, dateTime.toLocalDate())) {
+        is SkylightDay.AlwaysDaytime -> true
+        is SkylightDay.NeverLight -> false
+        is SkylightDay.Typical -> skylightDay.isLightAt(dateTime)
+    }
+
+/**
+ * Determine whether it is dark outside at the given [coordinates] at the given [dateTime], where "dark" means before
+ * dawn or after dusk on the given date.
+ */
+fun Skylight.isDark(coordinates: Coordinates, dateTime: ZonedDateTime) = !isLight(coordinates, dateTime)
+//endregion
+
+//region SkylightForCoordinates
+/**
+ * @param dateTime The date-time at which to check for lightness.
+ * @return Whether it is light outside at the [SkylightForCoordinates]'s coordinates at the given date-time, where
+ * "light" means after dawn and before dusk on the given date.
+ */
+fun SkylightForCoordinates.isLight(dateTime: ZonedDateTime) = skylight.isLight(coordinates, dateTime)
+
+/**
+ * @param dateTime The date-time at which to check for darkness.
+ * @return Whether it is dark outside at the [SkylightForCoordinates]'s coordinates at the given date-time, where "dark"
+ * means before dawn or after dusk on the given date.
+ */
+fun SkylightForCoordinates.isDark(dateTime: ZonedDateTime) = skylight.isDark(coordinates, dateTime)
+//endregion
+
+private fun SkylightDay.Typical.isLightAt(time: ZonedDateTime) =
+    dawn == null || (dawn.isBefore(time) && (dusk == null || dusk.isAfter(time)))
