+++ skylight/src/main/java/drewhamilton/skylight/SkylightForCoordinates.kt
@@ -1,15 +1,11 @@
 package drewhamilton.skylight
 
 import java.time.LocalDate
-import java.time.ZonedDateTime
-import javax.inject.Inject
 
 /**
  * A convenient [Skylight] wrapper for a single location.
  */
-// TODO: Determine if @Inject internal constructor works with Dagger in another module
-// TODO: Investigate assisted injection
-class SkylightForCoordinates @Inject internal constructor(
+class SkylightForCoordinates internal constructor(
     internal val skylight: Skylight,
     val coordinates: Coordinates
 ) {
@@ -26,17 +22,3 @@ class SkylightForCoordinates @Inject internal constructor(
  * @return a [SkylightForCoordinates] wrapping the given [Skylight] and [Coordinates]
  */
 fun Skylight.forCoordinates(coordinates: Coordinates) = SkylightForCoordinates(this, coordinates)
-
-/**
- * @param dateTime The date-time at which to check for lightness.
- * @return Whether it is light outside at the [SkylightForCoordinates]'s coordinates at the given date-time, where
- * "light" means after dawn and before dusk on the given date.
- */
-fun SkylightForCoordinates.isLight(dateTime: ZonedDateTime) = skylight.isLight(coordinates, dateTime)
-
-/**
- * @param dateTime The date-time at which to check for darkness.
- * @return Whether it is dark outside at the [SkylightForCoordinates]'s coordinates at the given date-time, where "dark"
- * means before dawn or after dusk on the given date.
- */
-fun SkylightForCoordinates.isDark(dateTime: ZonedDateTime) = skylight.isDark(coordinates, dateTime)
