+++ skylight/src/main/java/drewhamilton/skylight/SkylightForDate.kt
@@ -1,14 +1,11 @@
 package drewhamilton.skylight
 
 import java.time.LocalDate
-import javax.inject.Inject
 
 /**
  * A convenient [Skylight] wrapper for a single date.
  */
-// TODO: Determine if @Inject internal constructor works with Dagger in another module
-// TODO: Investigate assisted injection
-class SkylightForDate @Inject internal constructor(
+class SkylightForDate internal constructor(
     private val skylight: Skylight,
     val date: LocalDate
 ) {
